Fix static method. Fix unit test asset path.
